Project ---eCommarce1/requirementnts.txt
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==2.1.12
django-allauth==0.42.0
django-ckeditor==5.9.0
django-countries==6.1.2
django-crispy-forms==1.9.2
django-js-asset==1.2.2
django-simple-history==2.11.0
idna==2.10
oauthlib==3.1.0
Pillow==7.2.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
stripe==2.49.0
urllib3==1.25.9
